+++ Assets/BulkTools/BulkToolsManager/Editor/BulkToolsManagementWindow.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 using UnityEditor;
 using UnityEditor.PackageManager;
 
-namespace BTools.Management.Editor
+namespace BTools.Management.EditorScripts
 {
     public class BulkToolsManagementWindow : EditorWindow
     {
@@ -22,7 +22,7 @@ namespace BTools.Management.Editor
         private static List<bool> moduleEnabled = new List<bool>();
         private static List<bool> platformsShown = new List<bool>();
         private static List<List<bool>> enabledPlatforms = new List<List<bool>>();
-        private Vector2 scrollPos = Vector2.zero;
+        private static Vector2 scrollPos = Vector2.zero;
 
 
         [MenuItem("BulkTools/Module Manager", priority = 0)]
@@ -34,12 +34,18 @@ namespace BTools.Management.Editor
 
         public void OnGUI()
         {
+            DrawGUI();
+        }
+
+        public static void DrawGUI()
+        {
             scrollPos = EditorGUILayout.BeginScrollView(scrollPos);
+            EditorGUILayout.Space();
             //Window description
             EditorGUILayout.LabelField("This is the module manager. Use this window to pick what modules you want enabled and on what platform.", EditorStyles.wordWrappedLabel);
 
+            EditorGUILayout.LabelField("Enabled Modules", EditorStyles.boldLabel);
             ApplyRevertButtons();
-
             if (!ready)
             {
                 DrawLine(horizontal: true);
@@ -55,7 +61,6 @@ namespace BTools.Management.Editor
                     DrawLine(horizontal: true);
                 }
             }
-
             ApplyRevertButtons();
             EditorGUILayout.EndScrollView();
         }
@@ -63,7 +68,7 @@ namespace BTools.Management.Editor
         /// <summary>
         /// Draws the apply and revert/refresh buttons in a horizontal group.
         /// </summary>
-        private void ApplyRevertButtons() 
+        private static void ApplyRevertButtons() 
         {
             EditorGUILayout.Space(5);
             GUILayout.BeginHorizontal();
@@ -85,7 +90,7 @@ namespace BTools.Management.Editor
         /// <summary>
         /// Draws a 1 unit seprator line
         /// </summary>
-        private void DrawLine(bool horizontal = true) 
+        private static void DrawLine(bool horizontal = true) 
         {
             if(horizontal) 
             {
@@ -97,7 +102,7 @@ namespace BTools.Management.Editor
             }
         }
 
-        private void DrawModuleGUI(int moduleIndex) 
+        private static void DrawModuleGUI(int moduleIndex) 
         {
             var module = modules[moduleIndex];
 
@@ -113,7 +118,7 @@ namespace BTools.Management.Editor
             }
 
             //Module Title foldout. Exit early if foldout closed
-            modulesShown[moduleIndex] = EditorGUILayout.Foldout(modulesShown[moduleIndex], titleContent);
+            modulesShown[moduleIndex] = EditorGUILayout.Foldout(modulesShown[moduleIndex], titleContent, true);
             EditorGUI.indentLevel++;
             if (!modulesShown[moduleIndex]) 
             {
@@ -129,26 +134,44 @@ namespace BTools.Management.Editor
             EditorGUI.BeginChangeCheck();
 
             //Module enabled in project state:
-            EditorGUI.BeginDisabledGroup(dependedOn);//If this module is depended on it can't be disabled
-            moduleEnabled[moduleIndex] = EditorGUILayout.Toggle("Enabled in Project", moduleEnabled[moduleIndex]);
-            EditorGUI.EndDisabledGroup();
+            if (module.normalAssembly.Length > 0 || module.editorAssembly.Length > 0)//If the module has any assembly attactched
+            {
+                EditorGUI.BeginDisabledGroup(dependedOn);//If this module is depended on it can't be disabled
+                var content = new GUIContent("Enabled in Project", dependedOn ? "Module is depended on, and can't be disabled." : "");
+                moduleEnabled[moduleIndex] = EditorGUILayout.Toggle(content, moduleEnabled[moduleIndex]);
+                EditorGUI.EndDisabledGroup();
+            }
+            else //The module has no assembly attached and can't be disabled because of it
+            {
+                EditorGUI.BeginDisabledGroup(true);
+                var content = new GUIContent("Enabled in Project", "Module has no assembly defined, and can't be disabled.");
+                EditorGUILayout.Toggle(content, true);
+                EditorGUI.EndDisabledGroup();
+            }
 
             //Platform specific enabled states:
-            platformsShown[moduleIndex] = EditorGUILayout.Foldout(platformsShown[moduleIndex], new GUIContent("Include on Platforms:", "Platforms that this module should be included on."));
-            EditorGUI.BeginDisabledGroup(!moduleEnabled[moduleIndex]);//If the whole module is disabled it's platforms can't be modified
-            if (platformsShown[moduleIndex]) 
+            if (module.normalAssembly.Length > 0)
             {
-                EditorGUI.indentLevel++;
-           
-                for(int i = 0; i < enabledPlatforms[moduleIndex].Count; i++) 
+                platformsShown[moduleIndex] = EditorGUILayout.Foldout(platformsShown[moduleIndex], new GUIContent("Include on Platforms:", "Platforms that this module should be included on."));
+                EditorGUI.BeginDisabledGroup(!moduleEnabled[moduleIndex]);//If the whole module is disabled it's platforms can't be modified
+                if (platformsShown[moduleIndex])
                 {
-                    EditorGUI.BeginDisabledGroup(platformsDependedOn.Contains(module.supportedPlatforms[i]));//If this module is depended on on this platform it can't be changed here.
-                    enabledPlatforms[moduleIndex][i] = EditorGUILayout.Toggle(new GUIContent(module.supportedPlatforms[i], " " + module.supportedPlatforms[i] + " "), enabledPlatforms[moduleIndex][i]);
-                    EditorGUI.EndDisabledGroup();
+                    EditorGUI.indentLevel++;
+
+                    for (int i = 0; i < enabledPlatforms[moduleIndex].Count; i++)
+                    {
+                        EditorGUI.BeginDisabledGroup(platformsDependedOn.Contains(module.supportedPlatforms[i]));//If this module is depended on on this platform it can't be changed here.
+                        enabledPlatforms[moduleIndex][i] = EditorGUILayout.Toggle(new GUIContent(module.supportedPlatforms[i], " " + module.supportedPlatforms[i] + " "), enabledPlatforms[moduleIndex][i]);
+                        EditorGUI.EndDisabledGroup();
+                    }
+                    EditorGUI.indentLevel--;
                 }
-                EditorGUI.indentLevel--;
+                EditorGUI.EndDisabledGroup();
             }
-            EditorGUI.EndDisabledGroup();
+            else 
+            {
+                EditorGUILayout.LabelField("Module does not include a runtime assembly, target platforms are irrelevant.");
+            }
 
             //Update dependencies if this module has been updated
             if (EditorGUI.EndChangeCheck() && moduleEnabled[moduleIndex]) 
@@ -175,7 +198,7 @@ namespace BTools.Management.Editor
         /// <param name="moduleIndex"></param>
         /// <param name="platformsRequired"></param>
         /// <returns></returns>
-        private bool CheckDependedOn(int moduleIndex, out List<string> platformsRequired) 
+        private static bool CheckDependedOn(int moduleIndex, out List<string> platformsRequired) 
         {
             bool dependedOn = false;
             platformsRequired = new List<string>();
@@ -203,7 +226,7 @@ namespace BTools.Management.Editor
         /// Updates the states of all dependencies recursively.
         /// </summary>
         /// <param name="moduleIndex"></param>
-        private void UpdateDependencies(int moduleIndex) 
+        private static void UpdateDependencies(int moduleIndex) 
         {
             if (!moduleEnabled[moduleIndex]) { return; } //If this module isn't enabled then it's dependencies don't matter exit early
             foreach (var dependency in modules[moduleIndex].dependencies) //Loop through the dependency names
@@ -233,7 +256,7 @@ namespace BTools.Management.Editor
         /// <summary>
         /// Reverts current settings to what is currently configured in the assemdefs by reinitializing the cache
         /// </summary>
-        private void RevertRefresh() 
+        private static void RevertRefresh() 
         {
             ready = false;
             EditorApplication.delayCall += UpdatePackageMode;
@@ -272,7 +295,7 @@ namespace BTools.Management.Editor
         {
             LoadAllPossibleAssemblyTargets();
             GetEnabledStates();
-
+            CompareToProjectSettings();
             ready = true;
         }
 
@@ -348,6 +371,50 @@ namespace BTools.Management.Editor
                 enabledPlatforms.Add(supportedPlatformsEnabledStates);
                 moduleEnabled.Add(enabledInEditor);
             }
+        }
+
+        private static void CompareToProjectSettings() 
+        {
+            var settings = BulkToolsConfig.LoadSettings();
+            foreach(var moduleConfig in settings) 
+            {
+                for (int moduleIndex = 0; moduleIndex < modules.Count; moduleIndex++) 
+                {
+                    if (modules[moduleIndex].name == moduleConfig.moduleName) 
+                    {
+                        moduleEnabled[moduleIndex] = moduleConfig.moduleEnabled;
+                        for(int platformIndex = 0; platformIndex < modules[moduleIndex].supportedPlatforms.Count; platformIndex++) 
+                        {
+                            enabledPlatforms[moduleIndex][platformIndex] = moduleConfig.enabledPlatforms.Contains(modules[moduleIndex].supportedPlatforms[platformIndex]);
+                        }
+                    }
+                }
+            }
+            ApplyEnabledStates();
+        }
+
+        private static void ApplyToProjectSettings() 
+        {
+            var settings = new List<ModuleConfig>();
+
+            for (int moduleIndex = 0; moduleIndex < modules.Count; moduleIndex++)
+            {
+                ModuleConfig moduleConfig = new ModuleConfig() 
+                {
+                    moduleName = modules[moduleIndex].name,
+                    moduleEnabled = moduleEnabled[moduleIndex],
+                    enabledPlatforms = new List<string>()
+                };
+                for (int platformIndex = 0; platformIndex < modules[moduleIndex].supportedPlatforms.Count; platformIndex++)
+                {
+                    if (enabledPlatforms[moduleIndex].Count > platformIndex && enabledPlatforms[moduleIndex][platformIndex]) 
+                    {
+                        moduleConfig.enabledPlatforms.Add(modules[moduleIndex].supportedPlatforms[platformIndex]); 
+                    }
+                }
+                settings.Add(moduleConfig);
+            }
+            BulkToolsConfig.SaveSettings(settings);
         }
 
         /// <summary>
@@ -438,6 +505,8 @@ namespace BTools.Management.Editor
                     }
                 }
             }
+            ApplyToProjectSettings();
+            AssetDatabase.SaveAssets();
             AssetDatabase.Refresh();
         }
     }