+++ Assets/BulkTools/MagicEvents/Editor/MagicEventsEditorWindow.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace BTools.MagicEvents.EditorScripts
+{
+    //TODO: When module manager supports assembly defines build this window to show all events being called and the methods they are calling.
+    public class MagicEventsEditorWindow
+    {
+
+    }
+}