+++ Assets/PackageBuildingTools/Editor/PackageBuildingTools.cs
@@ -14,32 +14,27 @@ public class PackageBuildingTools : MonoBehaviour
             folderPath = folderPath.Remove(folderPath.LastIndexOf('/'));
         }
         string path = System.IO.Path.Combine(folderPath, "New Module Meta.asset");
-        var newModule = new BTools.Management.Editor.ModuleMetaData();
+        var newModule = new BTools.Management.EditorScripts.ModuleMetaData();
         newModule.supportedPlatforms = new List<string>
         {
             "Android",
-            "EmbeddedLinux",
-            "GameCoreScarlett",
-            "GameCoreXboxOne",
             "iOS",
             "LinuxStandalone64",
-            "CloudRendering",
-            "macOSStandalone",
             "PS4",
             "PS5",
-            "QNX",
-            "Stadia",
             "Switch",
-            "tvOS",
-            "WSA",
-            "VisionOS",
             "WebGL",
-            "WindowsStandalone32",
             "WindowsStandalone64",
             "XboxOne"
         };
         AssetDatabase.CreateAsset(newModule, path);
         
         Debug.Log("Created new module meta at:" + path);
+    }
+
+    [MenuItem("Package Building Tools/Trigger Recompile")]
+    public static void TriggerRecompile() 
+    {
+        UnityEditor.Compilation.CompilationPipeline.RequestScriptCompilation();
     }
 }