+++ Assets/TestingTheScripts.cs
@@ -6,18 +6,28 @@ using BTools.Singletons;
 using BTools.StateMachines;
 using BTools.StaticTools;
 using BTools.UtilPack;
+using BTools.MoreAttributes;
+using BTools.MagicEvents; 
 
 public class TestingTheScripts : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
+{    
+    private void Awake()
     {
-        
+        MagicEvent.AddListener("", OnEvent);
+        MagicEvent.AddListener("RandomNumber", OnRandomNumber);
+
+        new MagicEvent("RandomNumber")
+            .AddData("number", Random.Range(0f, 1f))
+            .Invoke();
     }
 
-    // Update is called once per frame
-    void Update()
+    private void OnEvent(MagicEventContext context) 
     {
-        
+        Debug.Log(context.eventName);
+    }
+
+    private void OnRandomNumber(MagicEventContext context) 
+    {
+        Debug.Log(context.GetData<float>("number"));
     }
 }